pyproject.toml
[tool.poetry]
name = "rakt-food-truck"
version = "0.1.0"
description = ""
authors = ["xavi <xpeiro92@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "rakt_food_truck"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5"
pandas = "^2.1.3"
psycopg = "^3.1.14"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.5"
googlemaps = "^4.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
